requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
Django==4.1.2
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
fonttools==4.37.4
idna==3.4
inflection==0.5.1
jsonschema==4.16.0
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
